Move stream_id tests out
