+++ neqo-transport/src/connection/mod.rs
@@ -2572,44 +2572,6 @@ mod tests {
     }
 
     #[test]
-    fn bidi_stream_properties() {
-        let id1 = StreamIndex::new(4).to_stream_id(StreamType::BiDi, Role::Client);
-        assert_eq!(id1.is_bidi(), true);
-        assert_eq!(id1.is_uni(), false);
-        assert_eq!(id1.is_client_initiated(), true);
-        assert_eq!(id1.is_server_initiated(), false);
-        assert_eq!(id1.role(), Role::Client);
-        assert_eq!(id1.is_self_initiated(Role::Client), true);
-        assert_eq!(id1.is_self_initiated(Role::Server), false);
-        assert_eq!(id1.is_remote_initiated(Role::Client), false);
-        assert_eq!(id1.is_remote_initiated(Role::Server), true);
-        assert_eq!(id1.is_send_only(Role::Server), false);
-        assert_eq!(id1.is_send_only(Role::Client), false);
-        assert_eq!(id1.is_recv_only(Role::Server), false);
-        assert_eq!(id1.is_recv_only(Role::Client), false);
-        assert_eq!(id1.as_u64(), 16);
-    }
-
-    #[test]
-    fn uni_stream_properties() {
-        let id2 = StreamIndex::new(8).to_stream_id(StreamType::UniDi, Role::Server);
-        assert_eq!(id2.is_bidi(), false);
-        assert_eq!(id2.is_uni(), true);
-        assert_eq!(id2.is_client_initiated(), false);
-        assert_eq!(id2.is_server_initiated(), true);
-        assert_eq!(id2.role(), Role::Server);
-        assert_eq!(id2.is_self_initiated(Role::Client), false);
-        assert_eq!(id2.is_self_initiated(Role::Server), true);
-        assert_eq!(id2.is_remote_initiated(Role::Client), true);
-        assert_eq!(id2.is_remote_initiated(Role::Server), false);
-        assert_eq!(id2.is_send_only(Role::Server), true);
-        assert_eq!(id2.is_send_only(Role::Client), false);
-        assert_eq!(id2.is_recv_only(Role::Server), false);
-        assert_eq!(id2.is_recv_only(Role::Client), true);
-        assert_eq!(id2.as_u64(), 35);
-    }
-
-    #[test]
     fn test_conn_stream_create() {
         let mut client = default_client();
